Move some common logic to a SimpleResultSet implementation
