+++ scala-native-jdbc/src/main/scala/java/sql/ResultSet.scala
@@ -40,8 +40,14 @@ trait ResultSet extends AutoCloseable {
    */
   def getRow(): Int
 
+  /**
+   * Maps the given ResultSet column label to its ResultSet column index.
+   */
+  def findColumn(columnLabel: String): Int
+
   // Column value getters
   def getString(columnIndex: Int): String
+  def getShort(columnIndex: Int): Short
   def getInt(columnIndex: Int): Int
   def getLong(columnIndex: Int): Long
   def getDouble(columnIndex: Int): Double
@@ -52,6 +58,7 @@ trait ResultSet extends AutoCloseable {
 
   // Column value getters by column name
   def getString(columnLabel: String): String
+  def getShort(columnLabel: String): Short
   def getInt(columnLabel: String): Int
   def getLong(columnLabel: String): Long
   def getDouble(columnLabel: String): Double
